+++ src/components/calcite-accordion/calcite-accordion.stories.js
@@ -20,13 +20,13 @@ storiesOf("Accordion", module)
         ["multi", "single", "single-persist"],
         "multi"
       )}"
-      icon-type="${select("icon-type", ["chevron", "plus-minus"], "chevron")}"
+      icon-type="${select("icon-type", ["chevron", "caret", "plus-minus"], "chevron")}"
     >
     <calcite-accordion-item item-title="Accordion Item"><img src="http://placekitten.com/100/200" />
     </calcite-accordion-item>
     <calcite-accordion-item item-title="Accordion Item 2" active><img src="http://placekitten.com/100/200" />
     </calcite-accordion-item>
-    <calcite-accordion-item item-title="Accordion Item 3">
+    <calcite-accordion-item item-title="Accordion Item 3" item-subtitle="Something short about this item>
     <calcite-radio-group scale="s">
     <calcite-radio-group-item value="Yes" checked>Yes</calcite-radio-group-item>
     <calcite-radio-group-item value="No">No</calcite-radio-group-item>
@@ -53,13 +53,13 @@ storiesOf("Accordion", module)
         ["multi", "single", "single-persist"],
         "multi"
       )}"
-      icon-type="${select("icon-type", ["chevron", "plus-minus"], "chevron")}"
+      icon-type="${select("icon-type", ["chevron", "caret", "plus-minus"], "chevron")}"
     >
     <calcite-accordion-item item-title="Accordion Item"><img src="http://placekitten.com/100/200" />
     </calcite-accordion-item>
     <calcite-accordion-item item-title="Accordion Item 2" active><img src="http://placekitten.com/100/200" />
     </calcite-accordion-item>
-    <calcite-accordion-item item-title="Accordion Item 3">
+    <calcite-accordion-item item-title="Accordion Item 3" item-subtitle="Something short about this item>
     <calcite-radio-group scale="s">
     <calcite-radio-group-item value="Yes" checked>Yes</calcite-radio-group-item>
     <calcite-radio-group-item value="No">No</calcite-radio-group-item>
@@ -86,13 +86,13 @@ storiesOf("Accordion", module)
           ["multi", "single", "single-persist"],
           "multi"
         )}"
-        icon-type="${select("icon-type", ["chevron", "plus-minus"], "chevron")}"
+        icon-type="${select("icon-type", ["chevron", "caret", "plus-minus"], "chevron")}"
       >
       <calcite-accordion-item item-title="Accordion Item"><img src="http://placekitten.com/100/200" />
       </calcite-accordion-item>
       <calcite-accordion-item item-title="Accordion Item 2" active><img src="http://placekitten.com/100/200" />
       </calcite-accordion-item>
-      <calcite-accordion-item item-title="Accordion Item 3">
+      <calcite-accordion-item item-title="Accordion Item 3" item-subtitle="Something short about this item>
       <calcite-radio-group scale="s">
       <calcite-radio-group-item value="Yes" checked>Yes</calcite-radio-group-item>
       <calcite-radio-group-item value="No">No</calcite-radio-group-item>