+++ src/components/calcite-accordion/calcite-accordion.tsx
@@ -45,8 +45,10 @@ export class CalciteAccordion {
   @Prop({ mutable: true, reflect: true }) iconPosition: "start" | "end" = "end";
 
   /** specify the placement of the icon in the header, defaults to end */
-  @Prop({ mutable: true, reflect: true }) iconType: "chevron" | "plus-minus" =
-    "chevron";
+  @Prop({ mutable: true, reflect: true }) iconType:
+    | "chevron"
+    | "caret"
+    | "plus-minus" = "chevron";
 
   /** specify the selection mode - multi (allow any number of open items), single (allow one open item),
    * or single-persist (allow and require one open item), defaults to multi */
@@ -77,7 +79,7 @@ export class CalciteAccordion {
     let iconPosition = ["start", "end"];
     if (!iconPosition.includes(this.iconPosition)) this.iconPosition = "end";
 
-    let iconType = ["chevron", "plus-minus"];
+    let iconType = ["chevron", "caret", "plus-minus"];
     if (!iconType.includes(this.iconType)) this.iconType = "chevron";
 
     let theme = ["light", "dark"];